rename dev-tools to dev
